requirements.txt
Flask==3.0.3
psycopg[binary]==3.2.13
python-dotenv==1.0.1
gunicorn==21.2.0


